Inject Client dependency via configuration.
