+++ test/correios/cep_test.exs
@@ -17,48 +17,28 @@ defmodule Correios.CEPTest do
     reason: "CEP NAO ENCONTRADO"
   }
 
-  defmodule FakeClient do
-    def request("54250-610") do
-      response = """
-      <return>
-        <bairro>Cavaleiro</bairro>
-        <cep>54250610</cep>
-        <cidade>Jaboatão dos Guararapes</cidade>
-        <end>Rua Fernando Amorim</end>
-        <uf>PE</uf>
-      </return>
-      """
-
-      {:ok, response}
-    end
-
-    def request("00000-000") do
-      {:error, "<faultstring>CEP NAO ENCONTRADO</faultstring>"}
-    end
-  end
-
   describe "find_address/1 when zip code is found" do
     test "returns address" do
-      assert Subject.find_address("54250-610", FakeClient) == {:ok, @address}
+      assert Subject.find_address("54250-610") == {:ok, @address}
     end
   end
 
   describe "find_address/1 when zip code is not found" do
     test "returns error" do
-      assert Subject.find_address("00000-000", FakeClient) == {:error, @error}
+      assert Subject.find_address("00000-000") == {:error, @error}
     end
   end
 
   describe "find_address!/1 when zip code is found" do
     test "returns address" do
-      assert Subject.find_address!("54250-610", FakeClient) == @address
+      assert Subject.find_address!("54250-610") == @address
     end
   end
 
   describe "find_address!/1 when zip code is not found" do
     test "raises error" do
       assert_raise Error, "CEP NAO ENCONTRADO", fn ->
-        Subject.find_address!("00000-000", FakeClient)
+        Subject.find_address!("00000-000")
       end
     end
   end
